requirements.txt
absl-py==0.15.0
arabic-reshaper==2.1.3
asgiref==3.4.1
astunparse==1.6.3
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.7
clang==5.0
click==8.0.3
colorama==0.4.4
Django==3.2.6
django-filter==21.1
et-xmlfile==1.1.0
flatbuffers==1.12
ftfy==6.0.3
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==3.1.0a0
grpcio==1.41.1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
h5py==3.1.0
hpack==3.0.0
hstspreload==2021.11.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.4
mysql-connector-python==8.0.27
mysqlclient==2.0.3
nltk==3.6.5
numpy==1.19.5
oauthlib==3.1.1
openpyxl==3.0.9
opt-einsum==3.3.0
pandas==1.3.4
pickle-mixin==1.0.2
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.1
regex==2021.11.2
reportlab==3.6.2
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.1
six==1.15.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3
sqlparse==0.4.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
threadpoolctl==3.0.0
tqdm==4.62.3
tweepy==4.3.0
typing-extensions==3.7.4.3
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
